requirements.txt
transformers>=4.26.0
torch>=1.13.0
pillow>=9.0.0
datasets>=2.9.0
sentencepiece>=0.1.97
requests>=2.28.0
protobuf>=3.20.0
